requirements-client.txt
paho-mqtt
sgp4
numpy
pyserial
pynmea2
pigpio
tomlkit
